+++ src/components/ButtonLinkTimao/index.jsx
@@ -1,0 +1,6 @@
+import { Link } from 'react-router';
+import ButtonTimao from '../ButtonTimao';
+
+const ButtonLinkTimao = ButtonTimao.withComponent(Link);
+
+export default ButtonLinkTimao;