+++ src/index.jsx
@@ -1,13 +1,17 @@
 import React from 'react';
 import { createRoot } from 'react-dom/client';
+import { BrowserRouter } from 'react-router';
 import App from './App.jsx';
 
 import './styles/settings/colours.css';
 import './styles/reset.css'
 
 const root = createRoot(document.getElementById('root'));
-root.render(<React.StrictMode>
-    <App />
+root.render(
+  <React.StrictMode>
+    <BrowserRouter>
+      <App />
+    </BrowserRouter>
   </React.StrictMode>,
   document.getElementById('root')
 );