+++ src/pages/NovoVideo/index.jsx
@@ -1,0 +1,27 @@
+import styled from 'styled-components';
+import HighlightTimao from '../../components/HighlightTimao';
+
+const NovoVideoSection = styled.section`
+    flex: 1;
+    padding: 40px;
+    background-color: var(--color-black-dark);
+    color: var(--color-gray-light);
+`;
+
+const Title = styled.h1`
+    font-size: 32rem;
+    margin-bottom: 12rem;
+`;
+
+function NovoVideo() {
+  return (
+    <NovoVideoSection>
+      <Title>Novo vídeo</Title>
+      <p>
+        Cadastro de vídeos <HighlightTimao>em breve</HighlightTimao>.
+      </p>
+    </NovoVideoSection>
+  );
+}
+
+export default NovoVideo;